requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
biopython
google-generativeai
reportlab==4.0.7
matplotlib
Pillow
python-dotenv